move project copy actions to its own controller
Conflicts:
	app/controllers/projects_controller.rb
