+++ app/controllers/projects_controller.rb
@@ -101,46 +101,6 @@ class ProjectsController < ApplicationController
 
   end
 
-  def copy
-    @source_project = @project
-    UserMailer.with_deliveries(params[:notifications] == '1') do
-      @project = Project.new
-      @project.safe_attributes = params[:project]
-      @project.enabled_module_names = params[:enabled_modules]
-      if validate_parent_id && @project.copy_associations(@source_project, :only => params[:only])
-        @project.set_allowed_parent!(params[:project]['parent_id']) if params[:project].has_key?('parent_id')
-        flash[:notice] = l(:notice_successful_create)
-        redirect_to :controller => '/projects', :action => 'settings', :id => @project
-      elsif !@project.valid?
-        # Project was created
-        # But some objects were not copied due to validation failures
-        # (eg. issues from disabled types)
-        # TODO: inform about that
-        redirect_to :back
-      end
-    end
-  rescue ActiveRecord::RecordNotFound
-    redirect_to :back
-  end
-
-  def copy_project
-    from = params[:coming_from].to_sym || :settings
-    @issue_custom_fields = WorkPackageCustomField.find(:all, :order => "#{CustomField.table_name}.position")
-    @types = Type.all
-    @root_projects = Project.find(:all,
-                                  :conditions => "parent_id IS NULL AND status = #{Project::STATUS_ACTIVE}",
-                                  :order => 'name')
-    @copy_project = Project.copy_attributes(@project)
-    if @copy_project
-      @copy_project.identifier = Project.next_identifier if Setting.sequential_project_identifiers?
-    else
-      redirect_to :back
-    end
-    render :action => :"copy_from_#{from}"
-  rescue ActiveRecord::RecordNotFound
-    redirect_to :back
-  end
-
   # Show @project
   def show
     @users_by_role = @project.users_by_role
@@ -262,21 +222,6 @@ private
     render_404
   end
 
-  # Validates parent_id param according to user's permissions
-  # TODO: move it to Project model in a validation that depends on User.current
-  def validate_parent_id
-    return true if User.current.admin?
-    parent_id = params[:project] && params[:project][:parent_id]
-    if parent_id || @project.new_record?
-      parent = parent_id.blank? ? nil : Project.find_by_id(parent_id.to_i)
-      unless @project.allowed_parents.include?(parent)
-        @project.errors.add :parent_id, :invalid
-        return false
-      end
-    end
-    true
-  end
-
   def jump_to_project_menu_item
     if params[:jump]
       # try to redirect to the requested menu item
@@ -331,4 +276,20 @@ private
                                                                     .select(:type_id)
                                                                     .uniq)
   end
+
+  # Validates parent_id param according to user's permissions
+  # TODO: move it to Project model in a validation that depends on User.current
+  def validate_parent_id
+    return true if User.current.admin?
+    parent_id = params[:project] && params[:project][:parent_id]
+    if parent_id || @project.new_record?
+      parent = parent_id.blank? ? nil : Project.find_by_id(parent_id.to_i)
+      unless @project.allowed_parents.include?(parent)
+        @project.errors.add :parent_id, :invalid
+        return false
+      end
+    end
+    true
+  end
+
 end
