+++ app/views/copy_projects/copy_from_admin.html.erb
@@ -30,9 +30,9 @@ See doc/COPYRIGHT.rdoc for more details.
 <h2><%=l(:label_project_new)%></h2>
 
 <%= labelled_tabular_form_for @copy_project, :url => { :action => "copy" } do |f| %>
-  <%= render :partial => 'form', :locals => { :f => f, :project => @copy_project, :renderTypes => true } %>
+  <%= render :partial => 'projects/form', :locals => { :f => f, :project => @copy_project, :renderTypes => true } %>
 
-  <%= render :partial => "projects/copy_settings/copy_associations", :locals => { :project => @project } %>
+  <%= render :partial => "copy_projects/copy_settings/copy_associations", :locals => { :project => @project } %>
 
   <%= submit_tag l(:button_copy) %>
 <% end %>
