+++ app/views/copy_projects/copy_settings/_block_checkbox.html.erb
