+++ app/views/copy_projects/copy_settings/_copy_associations.html.erb
@@ -30,33 +30,33 @@ See doc/COPYRIGHT.rdoc for more details.
 <fieldset class="copy box">
   <legend><%= l(:button_copy) %></legend>
 
-  <%= render :partial => "projects/copy_settings/block_checkbox",
+  <%= render :partial => "copy_projects/copy_settings/block_checkbox",
              :locals => { :name => "queries", :checked => true, :label => l(:label_query_plural),
                           :count => project.queries.count } %>
-  <%= render :partial => "projects/copy_settings/block_checkbox",
+  <%= render :partial => "copy_projects/copy_settings/block_checkbox",
              :locals => { :name => "boards", :checked => false, :label => l(:label_board_plural),
                           :count => project.boards.count } %>
-  <%= render :partial => "projects/copy_settings/block_checkbox",
+  <%= render :partial => "copy_projects/copy_settings/block_checkbox",
              :locals => { :name => "members", :checked => true, :label => l(:label_member_plural),
                           :count => project.members.count } %>
-  <%= render :partial => "projects/copy_settings/block_checkbox",
+  <%= render :partial => "copy_projects/copy_settings/block_checkbox",
              :locals => { :name => "project_associations",
                           :checked => true,
                           :label => l(:'timelines.project_menu.project_associations'),
                           :count => project.project_associations.count } %>
-  <%= render :partial => "projects/copy_settings/block_checkbox",
+  <%= render :partial => "copy_projects/copy_settings/block_checkbox",
              :locals => { :name => "timelines", :checked => true, :label => l(:'timelines.timelines'),
                           :count => project.timelines.count } %>
-  <%= render :partial => "projects/copy_settings/block_checkbox",
+  <%= render :partial => "copy_projects/copy_settings/block_checkbox",
              :locals => { :name => "versions", :checked => true, :label => l(:label_version_plural),
                           :count => project.versions.count } %>
-  <%= render :partial => "projects/copy_settings/block_checkbox",
+  <%= render :partial => "copy_projects/copy_settings/block_checkbox",
              :locals => { :name => "wiki", :checked => true, :label => l(:label_wiki_page_plural),
                           :count => project.wiki.nil? ? 0 : project.wiki.pages.count } %>
-  <%= render :partial => "projects/copy_settings/block_checkbox",
+  <%= render :partial => "copy_projects/copy_settings/block_checkbox",
              :locals => { :name => "work_packages", :checked => false, :label => l(:label_work_package_plural),
                           :count => project.work_packages.count } %>
-  <%= render :partial => "projects/copy_settings/block_checkbox",
+  <%= render :partial => "copy_projects/copy_settings/block_checkbox",
              :locals => { :name => "categories", :checked => true,
                           :label => l(:label_work_package_category_plural),
                           :count => project.categories.count } %>
