+++ app/views/copy_projects/copy_settings/_hidden_custom_fields.html.erb
