+++ app/views/copy_projects/copy_settings/_hidden_types.html.erb
